ch4codexv1.1/requirements.txt
numpy
scipy
pandas
matplotlib
cvxpy
pyyaml
